+++ x/relationships/keeper/handler_test.go
@@ -14,28 +14,31 @@ func (suite *KeeperTestSuite) Test_handleMsgCreateRelationship() {
 	receiver, err := sdk.AccAddressFromBech32("cosmos1cjf97gpzwmaf30pzvaargfgr884mpp5ak8f7ns")
 	suite.NoError(err)
 
+	subspace := "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e"
+
 	tests := []struct {
 		name                string
 		msg                 types.MsgCreateRelationship
-		storedRelationships []sdk.AccAddress
+		storedRelationships types.Relationships
 		expErr              error
 		expEvent            sdk.Event
 	}{
 		{
 			name:                "Relationship already created returns error",
-			msg:                 types.NewMsgCreateRelationship(sender, receiver),
-			storedRelationships: []sdk.AccAddress{receiver},
+			msg:                 types.NewMsgCreateRelationship(sender, receiver, subspace),
+			storedRelationships: types.Relationships{types.NewRelationship(receiver, subspace)},
 			expErr:              sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, fmt.Sprintf("relationship already exists with %s", receiver)),
 		},
 		{
 			name:                "Relationship has been saved correctly",
-			msg:                 types.NewMsgCreateRelationship(sender, receiver),
+			msg:                 types.NewMsgCreateRelationship(sender, receiver, subspace),
 			storedRelationships: nil,
 			expErr:              nil,
 			expEvent: sdk.NewEvent(
 				types.EventTypeRelationshipCreated,
 				sdk.NewAttribute(types.AttributeRelationshipSender, sender.String()),
 				sdk.NewAttribute(types.AttributeRelationshipReceiver, receiver.String()),
+				sdk.NewAttribute(types.AttributeRelationshipSubspace, subspace),
 			),
 		},
 	}
@@ -76,18 +79,24 @@ func (suite *KeeperTestSuite) Test_handleMsgDeleteRelationship() {
 	addr2, err := sdk.AccAddressFromBech32("cosmos1cjf97gpzwmaf30pzvaargfgr884mpp5ak8f7ns")
 	suite.NoError(err)
 
+	subspace := "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e"
+
 	store := suite.ctx.KVStore(suite.keeper.StoreKey)
 	store.Set(types.RelationshipsStoreKey(suite.testData.user),
-		suite.keeper.Cdc.MustMarshalBinaryBare(&[]sdk.AccAddress{addr1, addr2}))
+		suite.keeper.Cdc.MustMarshalBinaryBare(&types.Relationships{
+			types.NewRelationship(addr1, subspace),
+			types.NewRelationship(addr2, subspace),
+		}))
 
-	testMsg := types.NewMsgDeleteRelationship(suite.testData.user, addr1)
+	testMsg := types.NewMsgDeleteRelationship(suite.testData.user, addr1, subspace)
 
 	handler := keeper.NewHandler(suite.keeper)
 	res, err := handler(suite.ctx, testMsg)
 
 	suite.NoError(err)
 
-	suite.Equal([]sdk.AccAddress{addr2}, suite.keeper.GetUserRelationships(suite.ctx, suite.testData.user))
+	suite.Equal(types.Relationships{types.NewRelationship(addr2, subspace)},
+		suite.keeper.GetUserRelationships(suite.ctx, suite.testData.user))
 
 	// Check the events
 	suite.Len(res.Events, 1)
@@ -95,6 +104,7 @@ func (suite *KeeperTestSuite) Test_handleMsgDeleteRelationship() {
 		types.EventTypeRelationshipsDeleted,
 		sdk.NewAttribute(types.AttributeRelationshipSender, suite.testData.user.String()),
 		sdk.NewAttribute(types.AttributeRelationshipReceiver, addr1.String()),
+		sdk.NewAttribute(types.AttributeRelationshipSubspace, subspace),
 	))
 }