+++ x/relationships/types/events.go
@@ -8,6 +8,7 @@ const (
 	// Relationships attributes
 	AttributeRelationshipSender   = "relationship_sender"
 	AttributeRelationshipReceiver = "relationship_receiver"
+	AttributeRelationshipSubspace = "relationship_subspace"
 
 	// UserBlocks events
 	EventTypeBlockUser   = "block_user"