+++ x/relationships/types/genesis.go
@@ -3,17 +3,17 @@ package types
 import (
 	"fmt"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/desmos-labs/desmos/x/relationships/types/models"
 )
 
 // GenesisState contains the data of the genesis state for the profile module
 type GenesisState struct {
-	UsersRelationships map[string][]sdk.AccAddress `json:"users_relationships"`
+	UsersRelationships map[string]models.Relationships `json:"users_relationships"`
 	UsersBlocks        []UserBlock                 `json:"users_blocks"`
 }
 
 // NewGenesisState creates a new genesis state
-func NewGenesisState(usersRelationships map[string][]sdk.AccAddress, usersBlocks []UserBlock) GenesisState {
+func NewGenesisState(usersRelationships map[string]models.Relationships) GenesisState {
 	return GenesisState{
 		UsersRelationships: usersRelationships,
 		UsersBlocks:        usersBlocks,
@@ -23,7 +23,7 @@ func NewGenesisState(usersRelationships map[string][]sdk.AccAddress, usersBlocks
 // DefaultGenesisState returns a default GenesisState
 func DefaultGenesisState() GenesisState {
 	return GenesisState{
-		UsersRelationships: map[string][]sdk.AccAddress{},
+		UsersRelationships: map[string]models.Relationships{},
 		UsersBlocks:        []UserBlock{},
 	}
 }
@@ -31,9 +31,9 @@ func DefaultGenesisState() GenesisState {
 // ValidateGenesis validates the given genesis state and returns an error if something is invalid
 func ValidateGenesis(data GenesisState) error {
 	for _, relationships := range data.UsersRelationships {
-		for _, address := range relationships {
-			if address.Empty() {
-				return fmt.Errorf("invalid address %s", address)
+		for _, rel := range relationships {
+			if rel.Recipient.Empty() {
+				return fmt.Errorf("invalid relationship's recipient address %s", rel)
 			}
 		}
 	}