+++ x/reports/types/alias.go
@@ -22,10 +22,10 @@ var (
 	// functions aliases
 	NewMsgReportPost      = msgs.NewMsgReportPost
 	RegisterMessagesCodec = msgs.RegisterMessagesCodec
-	RegisterModelsCodec   = models.RegisterModelsCodec
 	ReportStoreKey        = models.ReportStoreKey
 	NewReportResponse     = models.NewReportResponse
 	NewReport             = models.NewReport
+	RegisterModelsCodec   = models.RegisterModelsCodec
 
 	// variable aliases
 	MsgsCodec              = msgs.MsgsCodec