+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Version 0.12.0
+## Changes
+- Changed `relationships`' implementation adding a `subspace` field to identify in which app
+  users make relationships (#266)
 - Implemented the possibility to (un)blocking a specific user (#169)
 
 # Version 0.11.0